pyproject.toml
[project]
name = "turo-crawler"
version = "0.1.0"
description = "A web crawler using Playwright and CDP to connect to Chrome browser instances"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.40.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.12.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]
